SDK: Remove server scheme autodetection
This has been deprecated in a499d2ec107612eb30c23740d5a02813ecba2cf4. For
maximum compatibility, keep supporting schemaless URLs, but always assume
HTTPS instead of autodetecting.
