+++ changelog.d/20260527_164323_roman_rm_scheme_autodetection.md
@@ -1,0 +1,5 @@
+### Removed
+
+- \[SDK, CLI\] Removed server schema autodetection, previously deprecated
+  in v2.36.0
+  (<https://github.com/cvat-ai/cvat/pull/10673>)
