+++ site/content/en/docs/api_sdk/sdk/highlevel-api.md
@@ -130,9 +130,9 @@ with Client("https://app.cvat.ai", config=config) as client:
 {{% alert title="Note" color="primary" %}}
 Historically, the SDK has allowed the URL scheme (`http:` or `https:`)
 to be omitted, and would attempt to automatically detect the protocol.
-This behavior is deprecated due to being inherently insecure,
-and will be removed in a future version.
-To avoid future breakage, make sure to specify the scheme explicitly.
+This automatic detection has been removed due to being inherently insecure.
+Now, if the scheme is omitted, the SDK assumes `https:`.
+For clarity, it is recommended to always specify the scheme explicitly.
 {{% /alert %}}
 
 When the server is located, its version is checked. If an unsupported version is found,
