+++ tests/python/sdk/self-signed.crt
