+++ tests/python/sdk/self-signed.key
