+++ tests/python/sdk/util.py
@@ -2,15 +2,23 @@
 #
 # SPDX-License-Identifier: MIT
 
+import contextlib
+import http.server
+import ssl
 import textwrap
-from collections.abc import Container
+import threading
+from collections.abc import Container, Generator
 from dataclasses import dataclass
 from pathlib import Path
+from typing import Any
 from urllib.parse import urlparse
 
 import pytest
+import requests
 from cvat_sdk.api_client.rest import RESTClientObject
 from cvat_sdk.core.helpers import DeferredTqdmProgressReporter
+
+from shared.utils.config import BASE_URL
 
 
 def make_pbar(file, **kwargs):
@@ -109,3 +117,50 @@ class OrgResourceHierarchy:
     job_id: int
     issue_id: int | None = None
     comment_id: int | None = None
+
+
+@contextlib.contextmanager
+def https_reverse_proxy() -> Generator[str, None, None]:
+    ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
+    ssl_context.minimum_version = ssl.TLSVersion.TLSv1_2
+    cert_dir = Path(__file__).parent
+    ssl_context.load_cert_chain(cert_dir / "self-signed.crt", cert_dir / "self-signed.key")
+
+    with http.server.HTTPServer(("localhost", 0), _ProxyHttpRequestHandler) as proxy_server:
+        proxy_server.socket = ssl_context.wrap_socket(
+            proxy_server.socket,
+            server_side=True,
+        )
+        server_thread = threading.Thread(target=proxy_server.serve_forever)
+        server_thread.start()
+        try:
+            yield f"https://localhost:{proxy_server.server_port}"
+        finally:
+            proxy_server.shutdown()
+            server_thread.join()
+
+
+class _ProxyHttpRequestHandler(http.server.BaseHTTPRequestHandler):
+    def do_GET(self):
+        response = requests.get(**self._shared_request_args())
+        self._translate_response(response)
+
+    def do_POST(self):
+        body_length = int(self.headers["Content-Length"])
+
+        response = requests.post(data=self.rfile.read(body_length), **self._shared_request_args())
+        self._translate_response(response)
+
+    def _shared_request_args(self) -> dict[str, Any]:
+        headers = {k.lower(): v for k, v in self.headers.items()}
+        del headers["host"]
+
+        return {"url": BASE_URL + self.path, "headers": headers, "timeout": 60, "stream": True}
+
+    def _translate_response(self, response: requests.Response) -> None:
+        self.send_response(response.status_code)
+        for key, value in response.headers.items():
+            self.send_header(key, value)
+        self.end_headers()
+        # Need to use raw here to prevent requests from handling Content-Encoding.
+        self.wfile.write(response.raw.read())
